Move some Exif functions to apps/shared/avifexif
So that the avif_apps target does not depend on avif_internal.
